+++ lib/src/services/keyboard_service.dart
@@ -4,19 +4,20 @@ import 'package:flutter/services.dart';
 class KeyboardService {
   static bool isVisible(BuildContext context) {
     assert(context != null);
-    BuildContext myContext =
+    var myContext =
         context.findRootAncestorStateOfType<ScaffoldState>()?.context;
-    if (myContext == null) {
-      myContext = context;
-    }
-    bool keyboardVisisble = false;
+    myContext ??= context;
+    var keyboardVisisble = false;
     if (MediaQuery.of(myContext).viewInsets.bottom > 0) {
       keyboardVisisble = true;
     }
     return keyboardVisisble;
   }
 
-  static dismiss() {
+  static void dismiss({BuildContext unfocus}) {
     SystemChannels.textInput.invokeMethod('TextInput.hide');
+    if (unfocus != null) {
+      FocusScope.of(unfocus).unfocus();
+    }
   }
 }