+++ CHANGELOG.md
@@ -1,2 +1,8 @@
+## [1.0.1] - 13 sep. 2020.
+* added optional unfocus context to KeyboardService dismiss.
+* using unfocus context in KeyboardAutoDismiss.
+* added example code to README.md.
+* code changes for analysis warnings.
+
 ## [1.0.0] - 30 jan. 2020.
 * initial release.